+++ src/ts/src/public/account/AccountApiCreateRequest.ts
@@ -1,0 +1,13 @@
+/**
+ * DTO send to create an account
+ *
+ * *Can be performed without authorization*
+ */
+export interface AccountApiCreateRequest {
+  email: string;
+  password: string;
+}
+
+export interface AccountApiCreateResponse {
+  /* Empty on purpose */
+}