+++ src/ts/src/public/account/AccountApiMigrateAuth0Request.ts
@@ -1,0 +1,10 @@
+import { WithToken } from "./dto";
+
+export interface AccountApiMigrateAuth0Request {
+  email: string;
+  password: string;
+}
+
+export interface AccountApiMigrateAuth0Response extends WithToken {
+  accountId: string;
+}