+++ src/ts/src/public/account/AccountApiPasswordChangeRequest.ts
@@ -1,0 +1,10 @@
+import { WithToken } from "./dto";
+
+export interface AccountApiPasswordChangeRequest {
+  password: string;
+  newPassword: string;
+}
+
+export interface AccountApiPasswordChangeResponse {
+  /* Empty on purpose */
+}