+++ src/ts/src/public/account/AccountApiSetupRequest.ts
@@ -1,9 +1,0 @@
-
-export interface AccountApiSetupRequest {
-  userDisplayName: string;
-  primaryWorkspaceName: string;
-}
-
-export interface AccountApiSetupResponse {
-  primaryWorkspaceId: string;
-}